# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Ticaret API",
        "description": "JWT kimlik doğrulama ile E-Ticaret uygulaması için API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Kategoriler"
                ],
                "summary": "Tüm kategorileri getir",
                "description": "Sistemdeki tüm kategorileri listeler",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Kategoriler başarıyla getirildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kategoriler"
                ],
                "summary": "Yeni kategori oluştur",
                "description": "Yeni bir kategori oluşturur (Sadece Admin)",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Kategori adı",
                                        "type": "string",
                                        "example": "Elektronik"
                                    },
                                    "description": {
                                        "description": "Kategori açıklaması",
                                        "type": "string",
                                        "example": "Elektronik cihazlar ve aksesuarlar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategori başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori başarıyla oluşturuldu"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "403": {
                        "description": "Yasak - Admin erişimi gerekli"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{id}": {
            "put": {
                "tags": [
                    "Kategoriler"
                ],
                "summary": "Kategori güncelle",
                "description": "Belirli bir kategoriyi günceller (Sadece Admin)",
                "operationId": "7dcf71887e585de7f881e9eb98d86126",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek kategorinin ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Kategori adı",
                                        "type": "string",
                                        "example": "Elektronik"
                                    },
                                    "description": {
                                        "description": "Kategori açıklaması",
                                        "type": "string",
                                        "example": "Elektronik cihazlar ve aksesuarlar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori başarıyla güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori başarıyla güncellendi"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Kategori bulunamadı"
                    },
                    "403": {
                        "description": "Yasak - Admin erişimi gerekli"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Kategoriler"
                ],
                "summary": "Kategori sil",
                "description": "Belirli bir kategoriyi siler (Sadece Admin)",
                "operationId": "9b61b65823f4a21f182506b34ddf805c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Silinecek kategorinin ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori başarıyla silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori başarıyla silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Kategori bulunamadı"
                    },
                    "403": {
                        "description": "Yasak - Admin erişimi gerekli"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login veya register endpoint'inden alınan JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}